moved ranges to module
