+++ include/main.hpp
@@ -1,10 +1,5 @@
 #pragma once
-#include <RosbotDrive.h>
-
-#include <memory_debug_message_info.hpp>
 #include <microros.hpp>
-#include <map>
-#include "MultiDistanceSensor.h"
 
 static UARTSerial microros_serial(RPI_SERIAL_TX, RPI_SERIAL_RX);
